# cpu: x64: update parallel section for brgemm matmul reorder