Fixes MFDNN-13029

Limits the threads number for small shapes for better performance.

Performance data:
<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/xuxinzen/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/xuxinzen/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl65
	{mso-number-format:"0\.0000";}
.xl66
	{mso-number-format:0%;}
-->
</style>
</head>

<body link="#0563C1" vlink="#954F72">


command | current(%-time) | main(%-time) | rls-v3.6(%-time) | main/current | rls-v3.6/current
-- | -- | -- | -- | -- | --
numactl   -C 24-27 -m 1 ./tests/benchdnn/benchdnn --mode=P --reorder --reset   --allow-enum-tags-only=0 --engine=cpu --sdt=f32 --ddt=f32 --stag=acdb   --dtag=abcd --strides=: --attr-scales= --attr-zero-points=   --attr-scratchpad=user 1x232x14x14 | 0.0134 | 0.0378 | 0.0134 | 282% | 100%
numactl   -C 24-27 -m 1 ./tests/benchdnn/benchdnn --mode=P --reorder --reset   --allow-enum-tags-only=0 --engine=cpu --sdt=f32 --ddt=f32 --stag=abcd   --dtag=acdb --strides=: --attr-scales= --attr-zero-points=   --attr-scratchpad=user 1x58x28x28 | 0.0134 | 0.0449 | 0.0134 | 335% | 100%
numactl   -C 24-27 -m 1 ./tests/benchdnn/benchdnn --mode=P --reorder --reset   --allow-enum-tags-only=0 --engine=cpu --sdt=f32 --ddt=f32 --stag=acdb   --dtag=abcd --strides=: --attr-scales= --attr-zero-points=   --attr-scratchpad=user 1x464x7x7 | 0.0134 | 0.0479 | 0.0134 | 356% | 100%



</body>

</html>