@xuxinzen 2025-01-23
make test
enable benchdnn_nightly
disable test_device_gpu